[CET-6660] Move methods to utils
