+++ src/components/ReportsPage/HeatmapPage/HeatmapUtils.ts
@@ -22,7 +22,7 @@ import {
   RuleOutcome,
   ScorecardScoreNextSteps,
 } from '../../../api/types';
-import { groupBy as _groupBy, flatten as _flatten, values, uniq, intersection, isEmpty } from 'lodash';
+import { groupBy as _groupBy, flatten as _flatten, values, uniq, intersection, isEmpty, flatten, isNil } from 'lodash';
 import { filterNotUndefined } from '../../../utils/collections';
 import { isApplicableRuleOutcome } from '../../../utils/ScorecardRules';
 import { HomepageEntity } from '../../../api/userInsightTypes';
@@ -171,6 +171,45 @@ export const groupScoresByHierarchies = (groupedScores: StringIndexable<Scorecar
 
   return hierarchyGroupedData;
 }
+
+export const findHierarchyItem = (
+  nodes: HierarchyNode[],
+  lastItemIdentifier: string,
+): HierarchyNode | undefined => {
+  let item = nodes.find((item: HierarchyNode) => {
+    return item.node.tag === lastItemIdentifier;
+  });
+
+  if (!item) {
+    for (const nodeItem of nodes) {
+      const childItem = findHierarchyItem(nodeItem.orderedChildren, lastItemIdentifier);
+      if (childItem) {
+        item = childItem;
+        break;
+      }
+    }
+  }
+
+  return item;
+};
+
+export const getAllHierarchyDescendants = (treeNode: HierarchyNode): HierarchyNode[] => {
+  const children = flatten(
+    treeNode.orderedChildren.map(childNode => getAllHierarchyDescendants(childNode)),
+  );
+
+  return [treeNode, ...children];
+};
+
+export const getAllHierarchyNodesFromTree = (
+  nodes?: HierarchyNode[],
+): HierarchyNode[] => {
+  if (isNil(nodes) || isEmpty(nodes)) {
+    return [];
+  }
+
+  return flatten(nodes.map(domainNode => getAllHierarchyDescendants(domainNode)));
+};
 
 export const catalogToRelationsByEntityId = (entitiesByTag: StringIndexable<HomepageEntity>) => {
   const ownerEmailByEntityId = {} as Record<string, string[]>;
